backend/requirements.txt
fastapi
uvicorn
spacy
pymupdf
python-docx
pymongo
python-multipart
